+++ src/main/java/firok/tiths/common/SoundEvents.java
@@ -6,12 +6,17 @@ import net.minecraft.util.SoundEvent;
 
 public class SoundEvents
 {
-	public static final SoundEvent record1=$("record1");
+	public static final SoundEvent recordTinkersEfforts=$(Keys.soundTinkersEfforts);
+	public static final SoundEvent recordTinkersWill=$(Keys.soundTinkersWill);
 
+	/**
+	 @see net.minecraft.util.SoundCategory
+	 */
 	public static SoundEvent $(String key)
 	{
-		SoundEvent ret=new SoundEvent(new ResourceLocation(TinkersThings.MOD_ID,key));
+		ResourceLocation resource=new ResourceLocation(TinkersThings.MOD_ID,key);
+		SoundEvent sound=new SoundEvent(resource);
 //		SoundEvent.REGISTRY.register(); // 貌似用不着注册就能用
-		return ret;
+		return sound;
 	}
 }